+++ crates/primitives/src/env.rs
@@ -259,7 +259,7 @@ pub struct CfgEnv {
     /// Chain ID is introduced EIP-155
     pub chain_id: u64,
     /// KZG Settings for point evaluation precompile. By default, this is loaded from the ethereum mainnet trusted setup.
-    #[cfg(feature = "c-kzg")]
+    #[cfg(feature = "kzg")]
     #[cfg_attr(feature = "serde", serde(skip))]
     pub kzg_settings: crate::kzg::EnvKzgSettings,
     /// Bytecode that is created with CREATE/CREATE2 is by default analysed and jumptable is created.
@@ -374,7 +374,7 @@ impl Default for CfgEnv {
             chain_id: 1,
             perf_analyse_created_bytecodes: AnalysisKind::default(),
             limit_contract_code_size: None,
-            #[cfg(feature = "c-kzg")]
+            #[cfg(feature = "kzg")]
             kzg_settings: crate::kzg::EnvKzgSettings::Default,
             #[cfg(feature = "memory_limit")]
             memory_limit: (1 << 32) - 1,