+++ crates/interpreter/CHANGELOG.md
@@ -51,7 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [3.1.0](https://github.com/bluealloy/revm/compare/revm-interpreter-v3.0.0...revm-interpreter-v3.1.0) - 2024-02-22
 
 ### Added
-- bump c-kzg, add portable feature, make it default ([#1106](https://github.com/bluealloy/revm/pull/1106))
+- ~~bump c-kzg, add portable feature, make it default ([#1106](https://github.com/bluealloy/revm/pull/1106))~~
 
 ### Fixed
 - replace tuple in sstore return with struct ([#1115](https://github.com/bluealloy/revm/pull/1115))